docker-letsencrypt: domain and gandi user are env variables.
Up to now the letsencrypt docker had the credentials
hardcoded at build time. Now we have the gandi credentials
as well as domain and www.domain passed as environmental
variables. If these env variables are not set the container
will fail.

The credentials are still recorded in git, but the truth is they
will not match the ones in production from now on.