Moved app tests to androidTests and run them in android emulator.
We will start to need to run the tests in an android
environment to be able to test certain scenarios. So
far the change is not perfect as the existing tests
do not have any dependency to android, yet.

To achieve a reproducible test infrastructure a docker container
was created with the service android-test. This container is
run on the TestAndroidTests test.

Useless Log.java unit test was not migrated and was just deleted.

Upgraded android tools as well as gradle definitions so that the
android build system itself launches the android emulator.