ThriftMessageHander: Check if the old connection is really disposed.
Before, when we had a connection of the same device ongoing
we marked the old connection for removal and carried on, leading
to possible races due to the fact the old connection's tear down
may not be instantaneous, making the whole effort a bit half hearted.

Now we try multiple times and wait for the old connection to be fully
torn down before proceeding. I could not find a way to make the
connection handoff observable unfortunately :(