DeviceConnectionAPI: Accept and log when closing device connection fails.
There are situations, which should not happen, where the
update of the device connection on connection close fails.
For example when the device is removed while the connection
is ongoing, would lead to failure. Other parts of the program
are responsible for this condition not happening, but if they
fail then at least now we have some more info and reporting.