DeviceAPI: Do not allow device delete while it is in use.
Before we had funny situations where we could be deleting
devices that still had active connections. This could lead
to very undefined behavior, specially due to the fact that
deleting the device can also delete associated device connections.

What happens when you delete a device and associated device
connections while the connections are ongoing....bad things
that should no longer be possible.

Added a test for this case and needed to expose the deleteDevice
method in HttpClientHelper to the IntegrationTests.