MediaObjectAndTagAPI: LocalStorageHelper get preffix replaced with calculate.
In some of the methods with get prefix it implied we were actually
retrieving data from existing files when this was not the case. We
were calculating what it would be. This distinction will be important
when we use the LocalStorageHelper to retrieve data that is actually
real and not a calculation.