DroHub Gallery: Propagate preview and pictures code to the front end.
The gallery now shows pictures as well as their previews. This
required some refactoring where the action URLs that are needed
by the frontend are propagated through properties instead of
being hardcoded.

No test for this functionality yet unfortunately, but it should
already be possible to do them together with the upload media
functionality.