GalleryTimeLine: Added media type thumbnail presentation.
Quite incomplete as the designer differentiates between recording
and video but at this point we do not pass such data to the front
end. Not such a big deal yet because we do not a way to sync video
media from the drone to the server yet. When we do we need to add
this metadata. For now i am happy with the results