Gallery: Put tag search bar under main top bar.
Before it was a bit floating by itself. Now it is in a visually
recognizable section of the page.

The placeholder text is now correctly propagated and displayed.

We needed to add some configurability to the default margins of the Main
layout as well as to the top navbar, due to the special case where we
want a visually adjacent bar from the top.

Also cleaned up some dangling elements and commented out html code.