GalleryTimeLine: Corrected GalleryItemSelect.
Now we pass both the preview as well as the actual file
id to the item select component so that it can chose whether
the preview or the actual file is to be consider selected.

Basically if only the preview is available then the preview is
considered selected. Otherwise we always consider the actual
file selected.