DeviceConnectionAPI: Guarantee that each device connection has a directory.
Before this commit we were relying on Janus to create the connection
directory, but this is a supposition, that may not work out. For example
if janus is off, all the logic relying on the existence of the directory
will not work. Also Janus creates directories with it's own permissions,
which if ran from a docker means root. In testing and some scenarios this
means we cannot write to the connection dir due to permission denied.