ParrotMediaStore: onChange for video and photo listeners trigger even on first results.
Before we had a mechanism that did not provide the initial
media count, but only the delta after that initial. This is
a misleading contract, as this makes it impossible for us to
know the initial results. We put the filtering of the first
results only on the CopterHUDActivity as it is where it should
be.

ParrotMediaStore also now guarantees that the _video_items and
_photo_items have an initilialized list even if empty.