app and backend: Added support for MediaInfoCardFragment in the LobbyActivity.
On the lobby now we have a card that shows the pictures and videos
available in the drone if, we have a connection to one.

It also shows the pictures and videos available in the DroHub
backend. To achieve this we post process the GalleryModel that
was normally produced for the platform's web front end. Neat,
as this meant miminal work on the backend.