Simplify token provider.
We no longer rely on the default token generator from Asp.net as it is
kind of meant to be used in a different scenario, specifically as a
temporary token. WE do not really want that. For the app we want just
a way to authenticate the app once and then provide a token that allows
just the app to access the user's resources as an app. For this we just
use the hash from the user as this is secure and reproduceable during
time.