IntegrationTest.cs: Add a unique prefix to ExcludeUser test.
When the subscription has no time left the websocket connection fails
and the tests that rely on it just abort unexpectedly not even
cleaning up the user or resources. This makes things more confusing
due to the fact that exclude user tests also have the same user
suffix and thus seemingly unrelated tests can fail in an unscrutable
cascade.