DroHub: Change DateTime to DateTimeOffset where possible.
DateTime is too tied with human calendars and less tied with
unequivocal time point. For that function DateTimeOffset seems
better and the support for the unix epoch time is evidence.

This removes the need to assume or provide the DateTime kind
property. In my opinion DateTime is a piece of shit. Maybe i
am not understanding it, but that is my opinion.