Gallery: Pass frontend urls through gallery-time-line attributes.
Before, we had some hardocoded urls in individual components
which made the management of endpoints painful and fragile
as the front end constants needed to be updated as well. Failure
led to broken functionality. Now we have a clearly defined interface
between the frontend and backend where the URLs are defined.