Gallery: Make a flight the content container.
Previously we had a day grouping multiple flight sessions
in the the gallery front end and this was confusing. Now
we have a proper separation. The model was re-architected
to match this new data layout.

Removed RepositoryGalleryItem.cs as it was orphaned and was
not in use for a long time.