GalleryTimeLine: Cross site verification Token is now a property.
Before, the token was inside the page's <script> tag,
which made it hard to mock as well as integrate with tests.
Now it is a property as others.

This complicated the automated test data preparation as
the token is a parameter that is short lived. For such
properties we added a new temporary-test-data.json as
an interface for front end testing where the short lived
properties live. The structure should be the same as the
normal test-data.json.