DroHub.Test: Split front end tests into FrontEndTests.cs
Before we had all the tests for frontend in one single invocation
of the container. Now we adapted the vue-test container to take an
argument on the command which for our case is the test name. With
this we were able to split the front end tests into the respective
facts.

Some refactoring was required where the testUpload was movedd to
TestServerFixture so it can be used by different test classes.

A small bug was fixed where we assumed the temp-test-data existed
which was not a good thing. Now we always remove the previous
temp file when we are about to run a new test.