DroHub.Tests: Added GalleryDeleteFilesModel.spec.js.
Needed to overhaul the test infrastructure to allow for multiple
files to be uploaded at the same time, and have a callback
when they are finished.

Also made the GalleryDeleteFilesModal have the special data
for removal properties, thus having the coupling more explicit
between the parent and the controller.

On the test side things are quite simple and we have device remove
test and an actual removal test. Other scenarios could probably
be added but for now it is enough,