DroHub: Video preview now jpeg sprite instead of webp.
webp cannot have it's loop stopped which makes it very useless
for static preview picture. A sprite is very easily handled on
the browser with css and on android with rects. Android and web
frontend usage of these preview sprites not implemented in this
commit.

Uploaded videos now have their preview generated automatically.
Also before starting the server the preview images are also
generated.