TestDeleteMediaObjects now works with empty database.
We always assumed the database had old data to take the
last elements but if our test was the only one using the media
database, after the delete was successful there would not be
any element and the Last() call would fail. Now we consider
success if the database is empty after we delete the files.