MediaObjectAndTagAPI: Rename generateVideoPreview to generateVideoPreviewBatchForDirectory.
The name was misleading as this method was not generatin a single specific
preview for a video file, but actually scanning the directory for previews
to generate. This was misleading and confusing due to other parts of
MediaObjectAndTagAPI actually not calling that code when they want to
generate a preview for a specific file.