Drohub Gallery: Pass gallery init data through property.
This is because it was an inconsistency, where everything was passed
from backend to vue front end with attributes -> property mappings
except for the initial data. I hope there is no issue with the encodings
of json, but this simplifies the code base and allows for testability
where the init data is actually generated dynamically as required by
the test.