MediaObjectAndTagAPI: Fix possible duplicated previews from different sources.
Now that we are generating preview images as well as sprite
images of videos it may happen they there is a collision of
names, leading to possible overwriting of one or the other,
depending on the upload order. To avoid this we now add a
video-prefix to the preview sprite image.