DeviceRepositoryController: Replaced GetPhoto and DownloadVideo.
We no longer have different end points to get videos or photos.
Instead we have DownloadFile endpoint that works for both media
types, and a GetPreview endpoint that is also suitable for both
types of media.

Currently things are not working fully as there is no frontend
for non preview pictures. This is why we let get-photo-url be
the GetPreview.

The same is true for video previews. There is no frontend functionality
for it.