Drohub.Tests: Added TestMultipleFilesWithSameTimestampNotAllowed.
Due to the way we are making our synthetic scenario of uploading
files, we are uploading multiple files with the same timestamp,
which given we use the timestamp for the file name, means we would
have collision. Fortunately the current code already does not allow
such collisions, but this was not tested and i was not sure the behavior
would be correct. With this test, we guarantee the behavior is correct
and gives us a spec of how it should work.

We needed to also modify test to make sure that we can indeed upload
multiple files sets separated just by timestamp.