DeviceRepositoryController: Added NonEmptyEnumerableValidation.
While making a full validation the frontend test for deleteMediaObjects
failed due to a previous change where the second file stopped being uploaded.
This led to the frontend test succeeding but the validation at the
backend that the file had indeed been removed was failing.

It turns out that the front end test was passing because there was no
error condition when the delete action was requested with no media ids
to delete. This led to wrongful success. With this patch we add a
validator that makes sure the request needs to have a non empty list
of items to delete, returning bad request if this is not the case.

Also added test on the IntegrationTest part, to make sure this does not
happen again. It is likely that other endpoint actions have the same bug,
although not the object of this commit.