DroHub: Do not expose preview paths to frontend.
We finally removed preview path and their processing from the frontend.
Now, when a preview is required, regardless of the media type, the GetPreview
endpoint is called with the media id. It is up to the backend to serve the
correct preview file.

Some of the changes led to removal of methods that no longer were needed,
specifically doesFileExist.