DroHub: fix stripPreviewPrefix.
When the passed file to stripPreviewPrefix was from a
live stream or from a video the rules for converting
from preview file path to actual media path were more
complicated than just removing the preview prefix. Specifically
the preview extension of a preview file is not a video
and the video- prefix that exists for live stream also
is not present.

This commit also adds the GetPreview because the bug was
detected there. Before this fix the included test would
fail.