AndroidApplicationController: Added CreateDeviceConnection endpoint.
With this commit we can now create ad-hoc device connections. This is useful for
example so that we can include media from sessions which did not happen on drohub.
Up to now this was not possible.

upload helpers were changed to use this new functionality when creating the mock
device connections, thus furthering the test coverage with that scenario. The normal
path where the device connection is created by a real connection of a drone is tested
as well, so nothing is left uncovered.