drohub: Attempt at having secrets loaded through environmental variables.
We had secrets like database credentials as well as keys in the
appSettings. We had both appSettings and appSettings.Development
and both were unfortunately tracked by version control. Just
removing secrets from appSettings was not enough as then we neeeded
a way to pass secrets between deployments. Environmental variables
are ideal as they can be configured in docker-compose files or passed
when launching the Drohub backend. We then can have an env file
with credentials for development, for when we run manually or have
docker pass the credentials when in deployment.

To achieve this we replaced appSettings' secrets with placeholders, that
must be overlayed by environmental variables. If the placeholders are found
it means the secrets are missing and Drohub will not start.

This commit also wipes the environmental variables after it consumes them so
that no other part of the application can read them.

Test container compositions were moved to their own docker-compose file. This
allows the release deployer to just need the main docker-compose file.