Drohub.Tests: Define docker collection
I was not aware that xunit ran tests in parallel for each test class.
This was leading to some very weird issues when all the tests were
launched at once. By setting all the test classes that use docker
in the same collection then we fall into the default behavior of
xunit where it does not parallelize between test collections.