MJRPostProcessor: Put intermediate results in a temporary directory.
Before this commit the results from the mjr to media container
were put in the same directory as the mjr files. These results
were not final and should have not been made visible in a
client's directory. Now only the result after the post processor
is visible in the caller's directory.