TestServerFixture: Moved up test out of docker setup.
Thw Wait closure provided by the Docker class is a bit magic and exceptions on
it can cause container disposal issues. So we move that logic out and also add
a max time of 60 seconds to wait for everything to be up until we give up.

Also removed ForceBuild, as I think it is not needed.