drohub: network_mode host is now only for janus.
Made all containers use the containerized network. This was required
because new versions of docker-compose do not allow the system we
previously had where ports were defined while in network_mode host.

This required many adjustments because inter-container communication
is now done through the docker subnet instead of the general localhost
address as was possible before.

janus being in host network also required some work, as letting containers
accessing host network required some magic extra_hosts to be declared.

There might be issues with the firewall not allowing docker containers
to access the host network. If this is the case then further work might
be required in making janus lose access to the host network.