DroHub.Test Ability to select the video storage volume
This is useful so we can configure a different volume for testing
than the one used for production. The reason we need this is that
we need to access the contents of the video content volume for testing
but that requires that we add special options so that we can make it
a specific bind mount volume.